reservation

+++ src/main/java/com/example/demo/department/DepartmentRepository.java
@@ -3,5 +3,5 @@ package com.example.demo.department;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface DepartmentRepository extends JpaRepository<Department, Long> {
-    Department findByDepartmentId(Long departmentId);
+    Department findByDepartmentId(Integer departmentId);
 }
